pyproject.toml
[project]
name = "template"
version = "0.1.0"
description = "A general purpose machine learning project template."
readme = "README.md"
authors = [{ name = "Will Rice", email = "wrice20@gmail.com" }]
requires-python = ">=3.12"
dependencies = [
  "git-python>=1.0.3",
  "lightning>=2.5.5",
  "mypy>=1.15.0",
  "pre-commit>=4.1.0",
  "pydantic>=2.12.2",
  "pydocstyle>=6.3.0",
  "pytest>=8.3.4",
  "python-dotenv>=1.1.1",
  "ruff>=0.9.7",
  "torch>=2.8.0",
  "torchaudio>=2.8.0",
  "torchvision>=0.23.0",
  "wandb>=0.22.2",
]

[build-system]
requires = ["uv_build>=0.9.2"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = ["template"]

[tool.ruff.lint.isort]
known-first-party = ["template"]

[tool.ruff.lint]
select = ["C", "E", "F", "I", "W", "D", "N", "B", "PTH", "ANN"]
ignore = ["D107"]
exclude = [".venv"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pytest.ini_options]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.mypy]
ignore_missing_imports = true
follow_imports_for_stubs = true
exclude = [".venv"]

[project.scripts]
train = "template.scripts.train:main"
